+++ OrangeDesignSystem/Sources/OrangeDesignSystem/Components/Cards/ODSCardHorizontal.swift
@@ -23,32 +23,37 @@
 
 import SwiftUI
 
-/// Model used to configure the `ODSCardTitleFirst` card.
-public struct ODSCardTitleFirstModel: Identifiable {
+/// Model used to configure the `ODSCardHorizontal` card.
+public struct ODSCardHorizontalModel: Identifiable {
     let title: String
     let subtitle: String?
-    let thumbnail: Image?
-    let imageSource: ODSImage.Source
     let supportingText: String?
+    let imageSource: ODSImage.Source
+    let imagePosition: ImagePosition
 
-    /// Initilize the model
+    public enum ImagePosition: Int {
+        case trailing = 0
+        case leading
+    }
+
+    /// Initialization
     ///
     /// - Parameters:
     ///  - title: The title to be displayed in the card.
-    ///  - thumbnail: The optional thumbnail: avatar, logo or icon.
     ///  - subtitle: Optional subtitle to be displayed in the card.
     ///  - imageSource: The image to be displayed in the card.
-    ///  - supportingText Optional text description to be displayed in the card.
+    ///  - imagePosition: The side where image is placed.
+    ///  - supportingText: Optional text description to be displayed in the card. The text displaying is limited to two lines (truncated tail).
     ///
     public init(title: String,
                 subtitle: String? = nil,
-                thumbnail: Image? = nil,
                 imageSource: ODSImage.Source,
+                imagePosition: ImagePosition = .leading,
                 supportingText: String? = nil) {
         self.title = title
         self.subtitle = subtitle
-        self.thumbnail = thumbnail
         self.imageSource = imageSource
+        self.imagePosition = imagePosition
         self.supportingText = supportingText
     }
 
@@ -61,22 +66,21 @@ public struct ODSCardTitleFirstModel: Identifiable {
 ///
 /// <a href="https://system.design.orange.com/0c1af118d/p/66bac5-cards/b/1591fb" target="_blank">ODS Card</a>.
 ///
-/// This is a full width card displayed with a title and a thumbnail on top as first element.
-/// This card is composed of three parts:
-/// - Header: with a title, an optinal subtitle and an optinal thmubnail
+/// This is a full width card displayed with an image on the left or on the right.
+/// This card is composed of two parts:
 /// - Media: (today an image)
-/// - Content: with an optinal supporting text and optional buttons (zero up to two)
+/// - Content: with a title, an optinal subtitle, an optinal supporting text and optional buttons (zero up to two)
 ///
-/// The card is configured using the model `ODSCardTitleFirstModel` and optional action buttons
+/// The card is configured using the model `ODSCardHorizontalModel` and optional action buttons
 /// can be provided through ViewBuilders `buttonContent1` and `buttonContent2`.
 ///
-/// Those view builder are usefull to provide buttons managed somewhere else to handle actions, manage disable state, apply style,...
+/// Those view builders are usefull to provide buttons managed somewhere else to handle actions, manage disable state, apply style,...
 ///
-public struct ODSCardTitleFirst<ButtonContent1, ButtonContent2>: View where ButtonContent1: View, ButtonContent2: View {
+public struct ODSCardHorizontal<ButtonContent1, ButtonContent2>: View where ButtonContent1: View, ButtonContent2: View {
 
+    private var model: ODSCardHorizontalModel
     private var buttonContent1: () -> ButtonContent1
     private var buttonContent2: () -> ButtonContent2
-    private let model: ODSCardTitleFirstModel
 
     /// Initialization with two buttons.
     ///
@@ -85,7 +89,7 @@ public struct ODSCardTitleFirst<ButtonContent1, ButtonContent2>: View where Butt
     ///  - buttonContent1: The button1 view builder
     ///  - buttonContent2: The button2 view builder
     ///
-    public init(model: ODSCardTitleFirstModel,
+    public init(model: ODSCardHorizontalModel,
                 @ViewBuilder buttonContent1: @escaping () -> ButtonContent1,
                 @ViewBuilder buttonContent2: @escaping () -> ButtonContent2)
     {
@@ -95,7 +99,7 @@ public struct ODSCardTitleFirst<ButtonContent1, ButtonContent2>: View where Butt
     }
 }
 
-extension ODSCardTitleFirst where ButtonContent2 == EmptyView {
+extension ODSCardHorizontal where ButtonContent2 == EmptyView {
 
     /// Initialization with one button.
     ///
@@ -103,7 +107,7 @@ extension ODSCardTitleFirst where ButtonContent2 == EmptyView {
     ///  - model: The model to configure the card.
     ///  - buttonContent1: The button1 view builder
     ///
-    public init(model: ODSCardTitleFirstModel,
+    public init(model: ODSCardHorizontalModel,
                 @ViewBuilder buttonContent1: @escaping () -> ButtonContent1)
     {
         self.model = model
@@ -112,13 +116,13 @@ extension ODSCardTitleFirst where ButtonContent2 == EmptyView {
     }
 }
 
-extension ODSCardTitleFirst where ButtonContent1 == EmptyView, ButtonContent2 == EmptyView {
+extension ODSCardHorizontal where ButtonContent1 == EmptyView, ButtonContent2 == EmptyView {
 
     /// Initialization without any button.
     ///
     /// - Parameter model: The model to configure the card.
     ///
-    public init(model: ODSCardTitleFirstModel) {
+    public init(model: ODSCardHorizontalModel) {
         self.model = model
         buttonContent1 = { EmptyView() }
         buttonContent2 = { EmptyView() }
@@ -127,89 +131,71 @@ extension ODSCardTitleFirst where ButtonContent1 == EmptyView, ButtonContent2 ==
 
 // MARK: View body implementation
 
-extension ODSCardTitleFirst {
-
+extension ODSCardHorizontal {
+    
     public var body: some View {
-        VStack(alignment: .leading, spacing: ODSSpacing.none) {
-            HStack(alignment: .center, spacing: ODSSpacing.s) {
-                model.thumbnail?
-                    .resizable()
-                    .aspectRatio(contentMode: .fill)
-                    .frame(width: 44.0, height: 44.0, alignment: .center)
-                    .clipShape(Circle())
-                    .accessibilityHidden(true)
 
-                VStack(alignment: .leading, spacing: ODSSpacing.none) {
-                    Text(model.title)
-                        .odsFont(.bodyBold)
-                        .frame(maxWidth: .infinity, alignment: .leading)
+        VStack(spacing: ODSSpacing.none) {
+            HStack(alignment: .center, spacing: ODSSpacing.none) {
+                Group {
+                    if case .leading = model.imagePosition {
+                        image
+                    }
 
-                    if let subtitle = model.subtitle, !subtitle.isEmpty {
-                        Text(subtitle)
-                            .odsFont(.bodyRegular)
+                    VStack(alignment: .leading, spacing: ODSSpacing.xs) {
+                        Text(model.title)
+                            .odsFont(.bodyBold)
+                            .frame(maxWidth: .infinity, alignment: .leading)
+
+                        if let subtitle = model.subtitle, !subtitle.isEmpty {
+                            Text(subtitle)
+                                .frame(maxWidth: .infinity, alignment: .leading)
+                        }
+                        if let supportingText = model.supportingText, !supportingText.isEmpty {
+                            Text(supportingText)
+                                .lineLimit(2)
+                                .frame(maxWidth: .infinity, alignment: .leading)
+                        }
+                    }
+                    .foregroundColor(.primary)
+                    .padding(.all, ODSSpacing.m)
+
+                    if case .trailing = model.imagePosition {
+                        image
                     }
                 }
             }
-            .multilineTextAlignment(.leading)
-            .foregroundColor(.primary)
-            .padding(.vertical, ODSSpacing.m)
+            .frame(minHeight: 128)
+
+            Divider()
+
+            // Add padding on buttons to avoid to have extra padding on
+            // HStack even if there are no view on buttons.
+            HStack(alignment: .center, spacing: ODSSpacing.m) {
+                buttonContent1().padding(.top, ODSSpacing.m)
+                    .padding(.bottom, ODSSpacing.m)
+                buttonContent2().padding(.top, ODSSpacing.m)
+                    .padding(.bottom, ODSSpacing.m)
+                Spacer()
+            }
             .padding(.horizontal, ODSSpacing.m)
-            .layoutPriority(100)
-
-            ODSImage(source: model.imageSource)
-                .aspectRatio(contentMode: .fill)
-                .accessibilityHidden(true)
-
-            VStack(alignment: .leading, spacing: ODSSpacing.none) {
-                if let supportingText = model.supportingText, !supportingText.isEmpty {
-                    Text(supportingText)
-                        .odsFont(.bodyRegular)
-                        .multilineTextAlignment(.leading)
-                        .padding(.horizontal, ODSSpacing.m)
-                        .padding(.top, ODSSpacing.s)
-                }
-
-                // Add padding on buttons to avoid to have extra padding on
-                // HStack even if there are no view on buttons.
-                HStack(spacing: ODSSpacing.m) {
-                    buttonContent1().padding(.top, ODSSpacing.m)
-                    buttonContent2().padding(.top, ODSSpacing.m)
-                }
-                .padding(.horizontal, ODSSpacing.m)
-            }
-            .modifier(PaddingModifier())
         }
-        .modifier(CardShadowModifier())
+        .background(ODSInternalColor.cardBackground.color)
+        .cornerRadius(10)
+        .shadow(radius: ODSSpacing.xs)
+        .padding(.all, ODSSpacing.s)
+    }
+
+    private var image: some View {
+        ODSImage(source: model.imageSource)
+            .accessibilityHidden(true)
+            .frame(width: 128)
+            .clipped()
     }
 }
 
-// MARK: Padding modifier
-
-private struct PaddingModifier: ViewModifier {
-    @State private var height: CGFloat?
-
-    private var bottomPadding: CGFloat {
-        height == 0 ? ODSSpacing.none : ODSSpacing.m
-    }
-
-    private var topPadding: CGFloat {
-        height == 0 ? ODSSpacing.none : ODSSpacing.xs
-    }
-
-    func body(content: Content) -> some View {
-        content
-            .readSize { size in
-                print("size\(size.height)")
-                height = size.height
-            }
-            .padding(.top, topPadding)
-            .padding(.bottom, bottomPadding)
-    }
-}
-
-// MARK: Previews
 #if DEBUG
-struct ODSCardTitleFirst_Previews: PreviewProvider {
+struct ODSCardHorizontal_Previews: PreviewProvider {
 
     struct Toast: View {
         @Binding var showText: String?
@@ -236,11 +222,11 @@ struct ODSCardTitleFirst_Previews: PreviewProvider {
         }
     }
 
-    static let model = ODSCardTitleFirstModel(
+    static let model = ODSCardHorizontalModel(
         title: ODSCCardPreviewData.title,
         subtitle: ODSCCardPreviewData.subtitle,
-        thumbnail: ODSCCardPreviewData.thumbnail,
         imageSource: .image(ODSCCardPreviewData.image),
+        imagePosition: .leading,
         supportingText: ODSCCardPreviewData.supportingText)
 
     struct TestView: View {
@@ -249,7 +235,7 @@ struct ODSCardTitleFirst_Previews: PreviewProvider {
 
         var body: some View {
             ScrollView {
-                ODSCardTitleFirst(model: ODSCardTitleFirst_Previews.model) {
+                ODSCardHorizontal(model: ODSCardHorizontal_Previews.model) {
                     ButtonAction(text: "Button 1") {
                         showTextInToast = "Button 1 Clicked"
                     }