+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unrelase](https://github.com/Orange-OpenSource/ods-ios/compare/0.10.0...qualif)
 
+- [DemoApp/SDK] Cards rename cards labels and components ([#360](https://github.com/Orange-OpenSource/ods-ios/issues/360)) 
 - [DemoApp] Use B&W Assets from InnovationCup Theme ([#317](https://github.com/Orange-OpenSource/ods-ios/issues/317))
 - [DemoApp] Update cards demo using realistic data ([#354](https://github.com/Orange-OpenSource/ods-ios/issues/354))
 - [DemoApp/SDK] Add SideBySide card component ([#56](https://github.com/Orange-OpenSource/ods-ios/issues/56))
@@ -49,7 +50,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - [SDK] Add NOTICE.txt file for Orange Images usage ([#295](https://github.com/Orange-OpenSource/ods-ios/issues/295))
 - [DemoApp] Update the Bar Tab variant list selection ([#260](https://github.com/Orange-OpenSource/ods-ios/issues/260))
-- [SDK] [Accessibility] Do not vocalize images in ODSSmallCard, ODSCardTitleFirst and ODSCardImageFirs (Bugs [#276](https://github.com/Orange-OpenSource/ods-ios/issues/276), [#275](https://github.com/Orange-OpenSource/ods-ios/issues/275), [#279](https://github.com/Orange-OpenSource/ods-ios/issues/279))
+- [SDK] [Accessibility] Do not vocalize images in ODSSmallCard, ODSCardTitleFirst and ODSCardImageFirst (Bugs [#276](https://github.com/Orange-OpenSource/ods-ios/issues/276), [#275](https://github.com/Orange-OpenSource/ods-ios/issues/275), [#279](https://github.com/Orange-OpenSource/ods-ios/issues/279))
 - [DemoApp] Update the Bar Navigation variant list selection ([#259](https://github.com/Orange-OpenSource/ods-ios/issues/259))  
 - [DemoApp/SDK] Add banners component ([#51](https://github.com/Orange-OpenSource/ods-ios/issues/51))
 - [SDK] Add documentation for bars-tab component ([#289](https://github.com/Orange-OpenSource/ods-ios/issues/289))